+++ integrations/unstructured/src/haystack_integrations/components/converters/unstructured/__init__.py
@@ -1,6 +1,6 @@
 # SPDX-FileCopyrightText: 2023-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-from unstructured_fileconverter_haystack.fileconverter import UnstructuredFileConverter
+from .converter import UnstructuredFileConverter
 
 __all__ = ["UnstructuredFileConverter"]
